Add 'tagFilters'.
